Relocate escape-related code to new files escape.hpp and escape.cpp.
