+++ components/files/configurationmanager.hpp
@@ -2,10 +2,8 @@
 #define COMPONENTS_FILES_CONFIGURATIONMANAGER_HPP
 
 #include <map>
-#include <queue>
 
 #include <boost/program_options.hpp>
-#include <boost/iostreams/filtering_stream.hpp>
 
 #include <components/files/fixedpath.hpp>
 #include <components/files/collections.hpp>
@@ -64,109 +62,6 @@ struct ConfigurationManager
 
         bool mSilent;
 };
-
-
-/**
- * \struct escape_hash_filter
- */
-struct escape_hash_filter : public boost::iostreams::input_filter
-{
-    static const int sEscape;
-    static const int sHashIdentifier;
-    static const int sEscapeIdentifier;
-
-    escape_hash_filter();
-    virtual ~escape_hash_filter();
-
-    template <typename Source> int get(Source & src);
-
-    private:
-        std::queue<int> mNext;
-        int mPrevious;
-
-        bool mSeenNonWhitespace;
-        bool mFinishLine;
-};
-
-struct unescape_hash_filter : public boost::iostreams::input_filter
-{
-    unescape_hash_filter();
-    virtual ~unescape_hash_filter();
-    
-    template <typename Source> int get(Source & src);
-
-    private:
-        bool expectingIdentifier;
-};
-
-template <typename Source>
-int unescape_hash_filter::get(Source & src)
-{
-    int character;
-    if (!expectingIdentifier)
-        character = boost::iostreams::get(src);
-    else
-    {
-        character = escape_hash_filter::sEscape;
-        expectingIdentifier = false;
-    }
-    if (character == escape_hash_filter::sEscape)
-    {
-        int nextChar = boost::iostreams::get(src);
-        int intended;
-        if (nextChar == escape_hash_filter::sEscapeIdentifier)
-            intended = escape_hash_filter::sEscape;
-        else if (nextChar == escape_hash_filter::sHashIdentifier)
-            intended = '#';
-        else if (nextChar == boost::iostreams::WOULD_BLOCK)
-        {
-            expectingIdentifier = true;
-            intended = nextChar;
-        }
-        else
-            intended = '?';
-        return intended;
-    }
-    else
-        return character;
-}
-
-/**
- * \class EscapeHashString
- */
-class EscapeHashString
-{
-    private:
-        std::string mData;
-    public:
-        static std::string processString(const std::string & str);
-
-        EscapeHashString();
-        EscapeHashString(const std::string & str);
-        EscapeHashString(const std::string & str, size_t pos, size_t len = std::string::npos);
-        EscapeHashString(const char * s);
-        EscapeHashString(const char * s, size_t n);
-        EscapeHashString(size_t n, char c);
-        template <class InputIterator>
-        EscapeHashString(InputIterator first, InputIterator last);
-
-        std::string toStdString() const;
-
-        friend std::ostream & operator<< (std::ostream & os, const EscapeHashString & eHS);
-};
-
-std::istream & operator>> (std::istream & is, EscapeHashString & eHS);
-
-struct EscapeStringVector
-{
-    std::vector<Files::EscapeHashString> mVector;
-
-    EscapeStringVector();
-    virtual ~EscapeStringVector();
-
-    std::vector<std::string> toStdStringVector() const;
-};
-
 } /* namespace Cfg */
 
 #endif /* COMPONENTS_FILES_CONFIGURATIONMANAGER_HPP */
